Ringo 138

+++ BackEnd/src/main/java/springwebsocket/webchat/friend/repository/JpaFriendshipRepository.java
@@ -5,6 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 import springwebsocket.webchat.friend.entity.Friendship;
+import springwebsocket.webchat.friend.repository.springdata.UserInfoMapping;
 import springwebsocket.webchat.member.entity.Member;
 import springwebsocket.webchat.friend.repository.springdata.SpringDataJpaFriendshipRepository;
 import springwebsocket.webchat.member.repository.springdata.SpringDataJpaMemberRepository;
@@ -25,13 +26,13 @@ public class JpaFriendshipRepository implements FriendshipRepository {
     private final SpringDataJpaMemberRepository memberRepository;
 
     @Override
-    public Friendship sendFriendRequest(Long senderId, String receiverEmail) {
+    public Friendship sendFriendRequest(String senderEmail, String receiverEmail) {
 
-        Optional<Member> senderMember = memberRepository.findById(senderId);
+        Optional<Member> senderMember = memberRepository.findByEmail(senderEmail);
         Optional<Member> receiverMember = memberRepository.findByEmail(receiverEmail);
 
         if (receiverMember.isPresent()) {
-            // 상대방 Id가 존재할 때
+            // 상대방 Email가 존재할 때
             Member sender = senderMember.get();
             Member receiver = receiverMember.get();
 
@@ -49,9 +50,9 @@ public class JpaFriendshipRepository implements FriendshipRepository {
     }
 
     @Override
-    public void acceptFriendRequestById(Long senderId, String receiverEmail) {
+    public String acceptFriendRequestById(String senderEmail, String receiverEmail) {
 
-        Optional<Member> senderMember = memberRepository.findById(senderId);
+        Optional<Member> senderMember = memberRepository.findByEmail(senderEmail);
         Optional<Member> receiverMember = memberRepository.findByEmail(receiverEmail);
 
 
@@ -61,47 +62,52 @@ public class JpaFriendshipRepository implements FriendshipRepository {
         // 기존의 Friendship 엔터티를 찾는다.
         Optional<Friendship> existingFriendship = friendshipRepository.findByFriendIdAndUserId(sender, receiver);
 
+        if(existingFriendship.isEmpty()) return "fail";
+
         existingFriendship.ifPresent(friendship -> {
             // Friendship 엔터티의 상태를 FRIENDS로 update.
             friendship.setStatus(FRIENDS);
             // 업데이트된 Friendship 엔터티를 저장.
 //            friendshipRepository.save(friendship);
         });
-
+        return "success";
     }
 
 
     @Override
-    public void rejectFriendRequestById(Long id, String Email) {
+    public String rejectFriendRequestById(String senderEmail, String receiverEmail) {
 
-        Optional<Member> memberMe = memberRepository.findById(id);
-        Optional<Member> memberYou = memberRepository.findByEmail(Email);
+        Optional<Member> memberMe = memberRepository.findByEmail(senderEmail);
+        Optional<Member> memberYou = memberRepository.findByEmail(receiverEmail);
 
         // 기존의 Friendship 엔터티를 찾는다.
         Optional<Friendship> existingFriendship = friendshipRepository.findByFriendIdAndUserId(memberMe.get(), memberYou.get());
+
+        if(existingFriendship.isEmpty()) return "fail";
 
         existingFriendship.ifPresent(friendship ->{
             // Friendship 엔터티를 삭제
             friendshipRepository.delete(existingFriendship.get());
         });
 
+        return "success";
     }
 
     @Override
-    public List<Member> findByFriendIdAndStatus(Long id) {
+    public List<UserInfoMapping> findByFriendIdAndStatus(Long id) {
         Optional<Member> userMember = memberRepository.findById(id);
         return friendshipRepository.findByFriendIdAndStatus(userMember.get(), Friendship.FriendshipStatus.PENDING);
     }
 
     @Override
-    public List<String> findByUserIdAndStatusOrFriendIdAndStatus(Long userId) {
+    public List<UserInfoMapping> findByUserIdAndStatusOrFriendIdAndStatus(Long userId) {
 
-        List<String> friendList = new ArrayList<>();
+        List<UserInfoMapping> friendList = new ArrayList<>();
 
         Optional<Member> userMember = memberRepository.findById(userId);
 
-        List<String> friendList1 = friendshipRepository.findFriendshipByUserIdAndStatus(userMember.get(), FRIENDS);
-        List<String> friendList2 = friendshipRepository.findFriendshipsByFriendIdAndStatus(userMember.get(), FRIENDS);
+        List<UserInfoMapping> friendList1 = friendshipRepository.findFriendshipByUserIdAndStatus(userMember.get(), FRIENDS);
+        List<UserInfoMapping> friendList2 = friendshipRepository.findFriendshipsByFriendIdAndStatus(userMember.get(), FRIENDS);
 
         friendList.addAll(friendList1);
         friendList.addAll(friendList2);
